Create account deletion information page
Added "account-deletion.html" document outlining the steps to delete an account, the data that will be removed, and the irreversible nature of the process. This ensures that users are fully informed about how to permanently delete their accounts and what data will be affected. Contact information for support is also included.